create useOpenCalendarEventInCommandMenu
