+++ packages/twenty-front/src/modules/command-menu/hooks/useCommandMenu.ts
@@ -1,7 +1,7 @@
 import { useRecoilCallback } from 'recoil';
 
 import { commandMenuSearchState } from '@/command-menu/states/commandMenuSearchState';
-import { IconCalendarEvent, IconDotsVertical, IconMail } from 'twenty-ui';
+import { IconDotsVertical, IconMail } from 'twenty-ui';
 
 import { useNavigateCommandMenu } from '@/command-menu/hooks/useNavigateCommandMenu';
 import { viewableRecordIdComponentState } from '@/command-menu/pages/record-page/states/viewableRecordIdComponentState';
@@ -51,54 +51,6 @@ export const useCommandMenu = () => {
         }
       },
     [closeCommandMenu, openCommandMenu],
-  );
-
-  const openCalendarEventInCommandMenu = useRecoilCallback(
-    ({ set }) => {
-      return (calendarEventId: string) => {
-        const pageComponentInstanceId = v4();
-
-        set(
-          viewableRecordIdComponentState.atomFamily({
-            instanceId: pageComponentInstanceId,
-          }),
-          calendarEventId,
-        );
-
-        // TODO: Uncomment this once we need to calendar event title in the navigation
-        // const objectMetadataItem = snapshot
-        //   .getLoadable(objectMetadataItemsState)
-        //   .getValue()
-        //   .find(
-        //     ({ nameSingular }) =>
-        //       nameSingular === CoreObjectNameSingular.CalendarEvent,
-        //   );
-
-        // set(
-        //   commandMenuNavigationMorphItemsState,
-        //   new Map([
-        //     ...snapshot
-        //       .getLoadable(commandMenuNavigationMorphItemsState)
-        //       .getValue(),
-        //     [
-        //       pageComponentInstanceId,
-        //       {
-        //         objectMetadataId: objectMetadataItem?.id,
-        //         recordId: calendarEventId,
-        //       },
-        //     ],
-        //   ]),
-        // );
-
-        navigateCommandMenu({
-          page: CommandMenuPages.ViewCalendarEvent,
-          pageTitle: 'Calendar Event',
-          pageIcon: IconCalendarEvent,
-          pageId: pageComponentInstanceId,
-        });
-      };
-    },
-    [navigateCommandMenu],
   );
 
   const openEmailThreadInCommandMenu = useRecoilCallback(
@@ -155,7 +107,6 @@ export const useCommandMenu = () => {
     onCommandMenuCloseAnimationComplete,
     navigateCommandMenu,
     toggleCommandMenu,
-    openCalendarEventInCommandMenu,
     openEmailThreadInCommandMenu,
   };
 };
